organizing calculus syntax, ticket #231. This commit might have side-effects for C++ users, since some functions move from global namespace to CasADi namespace.

+++ examples/cplusplus/det_minor_mx.cpp
@@ -139,7 +139,7 @@ int main() {
   // Some random matrix
   DMatrix A0(n*n,1,0);
   for(int i=0; i<n*n; ++i)
-    A0[i] = 0.1 + sqrt(i*1323 % 124);
+    A0[i] = 0.1 + sqrt(double(i*1323 % 124));
   cout << A0 << endl;
   
   detFun.setInput(A0);
